requirements.txt
Flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
requests>=2.28.0
